move conwstants from visorconfig to skyenv package ; small changes on skywire-cli rtfind to print help on no arguments or use the local / running visor's public key as the first argument if one argument is given
