+++ pkg/visor/visorconfig/values.go
@@ -10,11 +10,15 @@ import (
 	"os/exec"
 	"os/user"
 	"path/filepath"
+	"runtime"
 	"strings"
 	"time"
 
 	"github.com/bitfield/script"
+	"github.com/google/uuid"
+	"github.com/jaypipes/ghw"
 	"github.com/skycoin/dmsg/pkg/dmsg"
+	"github.com/zcalusic/sysinfo"
 
 	"github.com/skycoin/skywire-utilities/pkg/buildinfo"
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
@@ -22,6 +26,17 @@ import (
 )
 
 var (
+	//OS detection at runtime
+	OS = skyenv.OS
+	// SkywirePath is the path to the installation folder for the linux packages.
+	SkywirePath = skyenv.SkywirePath
+	// ConfigJSON is the config name generated by the skywire-autocofig script in the linux packages
+	ConfigJSON = skyenv.ConfigJSON
+	// SkyenvFilePath is the path to the SkyenvFile
+	SkyenvFilePath = skyenv.SkyenvFilePath
+	// SkyenvFile contains environmental variables which are detected by `skywire-autoconfig` / `skywire-cli config auto` to set default or persistent values
+	SkyenvFile = skyenv.SkyenvFile
+
 	// config file constants
 
 	// ConfigName is the default config name. Updated by setting config file path.
@@ -138,6 +153,11 @@ var (
 	RewardFile = skyenv.RewardFile
 )
 
+// SkywireConfig returns the full path to the package config
+func SkywireConfig() string {
+	return SkywirePath + "/" + ConfigJSON
+}
+
 // PkgConfig struct contains paths specific to the linux packages
 type PkgConfig struct {
 	LauncherBinPath string `json:"launcher"`
@@ -199,9 +219,9 @@ func HomePath() string {
 }
 
 // Config returns either UserConfig or PackageConfig based on permissions
-func Config() PkgConfig {
+func Config() skyenv.PkgConfig {
 	if IsRoot() {
-		return PackageConfig()
+		return skyenv.PackageConfig()
 	}
 	return UserConfig()
 }
@@ -304,3 +324,67 @@ var (
 	// VisorConfigFile will contain the path to the visor's config or `stdin` to denote that the config was read from STDIN
 	VisorConfigFile string
 )
+
+// Survey system hardware survey struct
+type Survey struct {
+	Timestamp      time.Time        `json:"timestamp"`
+	PubKey         cipher.PubKey    `json:"public_key,omitempty"`
+	SkycoinAddress string           `json:"skycoin_address,omitempty"`
+	GOOS           string           `json:"go_os,omitempty"`
+	GOARCH         string           `json:"go_arch,omitempty"`
+	SYSINFO        sysinfo.SysInfo  `json:"zcalusic_sysinfo,omitempty"`
+	IPInfo         *IPSkycoin       `json:"ip.skycoin.com,omitempty"`
+	IPAddr         *IPAddr          `json:"ip_addr,omitempty"`
+	Disks          *ghw.BlockInfo   `json:"ghw_blockinfo,omitempty"`
+	Product        *ghw.ProductInfo `json:"ghw_productinfo,omitempty"`
+	Memory         *ghw.MemoryInfo  `json:"ghw_memoryinfo,omitempty"`
+	UUID           uuid.UUID        `json:"uuid,omitempty"`
+	SkywireVersion string           `json:"skywire_version,omitempty"`
+}
+
+// SystemSurvey returns system survey
+func SystemSurvey() (Survey, error) {
+	var si sysinfo.SysInfo
+	si.GetSysInfo()
+	disks, err := ghw.Block()
+	if err != nil {
+		return Survey{}, err
+	}
+	product, err := ghw.Product()
+	if err != nil {
+		return Survey{}, err
+	}
+	memory, err := ghw.Memory()
+	if err != nil {
+		return Survey{}, err
+	}
+	for {
+		ipInfo := IPSkycoinFetch()
+		if ipInfo != nil {
+			break
+		}
+	}
+	s := Survey{
+		Timestamp:      time.Now(),
+		IPInfo:         IPSkycoinFetch(),
+		IPAddr:         IPA(),
+		GOOS:           runtime.GOOS,
+		GOARCH:         runtime.GOARCH,
+		SYSINFO:        si,
+		UUID:           uuid.New(),
+		Disks:          disks,
+		Product:        product,
+		Memory:         memory,
+		SkywireVersion: Version(),
+	}
+	return s, nil
+}
+
+func PackageConfig() skyenv.PkgConfig {
+	return skyenv.PackageConfig()
+}
+
+// UpdateCommand returns the commands which are run when the update button is clicked in the ui
+func UpdateCommand() []string {
+	return []string{`echo "not implemented"`}
+}
